Complete modularization of table.js.
