+++ src/AETable/eventListeners.js
@@ -1,0 +1,3 @@
+export function eventListeners(data, vars, settings) {
+
+}
