+++ src/controls/filters.js
@@ -1,0 +1,6 @@
+import { rate } from './filters/rate';
+import { custom } from './filters/custom';
+
+export const filters =
+    {rate: rate
+    ,custom: custom};
