+++ src/controls/filters/custom.js
@@ -1,0 +1,8 @@
+import { init } from './custom/init';
+import { get } from './custom/get';
+import { set } from './custom/set';
+
+export const custom =
+    {init: init
+    ,get: get
+    ,set: set};
