+++ src/controls/search.js
@@ -1,0 +1,10 @@
+import { init } from './search/init';
+import { get } from './search/get';
+import { set } from './search/set';
+import { clear } from './search/clear';
+
+export const search =
+    {init: init
+    ,get: get
+    ,set: set
+    ,clear: clear};
