+++ src/controls/search/get.js
@@ -1,0 +1,1 @@
+export function get() {}
