+++ src/eventListeners/customFilters.js
@@ -1,0 +1,8 @@
+export function customFilters(table, canvas, data, vars, settings) {
+    var filterCustom = canvas.selectAll(".custom-filters ul li select");
+
+  //redraw table without bootstrap multiselect
+    filterCustom.on('change', function() {
+        table.AETable.redraw(table, canvas, data, vars, settings);
+    });
+}
