+++ src/wrapper.js
@@ -1,5 +1,3 @@
-//import { table } from './table';
-import { util } from './util';
 import { init } from './init';
 import { colorScale } from './colorScale';
 import { layout } from './layout';
@@ -7,6 +5,7 @@ import { controls } from './controls';
 import { eventListeners } from './eventListeners';
 import { AETable } from './AETable';
 import { detailTable } from './detailTable';
+import { util } from './util';
 
 export default function aeTable() {
     const table =
